update axios test

+++ src/axios/axios.test.ts
@@ -18,6 +18,7 @@ import {
   deleteBandBookmark,
   patchArtistLink,
   patchBandLink,
+  postIssue,
 } from "./axios";
 import {
   IAlbum,
@@ -25,19 +26,17 @@ import {
   IArtistBookmark,
   IBand,
   IBandBookmark,
-  IHeaders,
   IMusic,
   IMusicBookmark,
-  IUser,
 } from "../interfaces";
 import testUser from "../constants/user.json";
 import routes from "../constants/routes.json";
 import { signUp } from "../ui/Form/SignUp";
 import { signIn } from "../ui/Form/SignIn";
-import { ITokenHeaders } from "../slices/currentUser/currentUser";
+import { setCurrentUser, setHeaders } from "../slices/currentUser/currentUser";
+import { store } from "../store";
 
-let headers: IHeaders;
-let user: IUser;
+const getCurrentUser = () => store.getState().currentUser.currentUser;
 
 // helpers
 const signUpTest = async () => {
@@ -49,8 +48,8 @@ const signUpTest = async () => {
     password: testUser.PASSWORD,
     password_confirmation: testUser.PASSWORD,
   });
-  headers = { headers: res.headers };
-  user = res.data.data;
+  store.dispatch(setHeaders(res.headers));
+  store.dispatch(setCurrentUser(res.data.data));
   return res;
 };
 const signInTest = async () => {
@@ -58,28 +57,11 @@ const signInTest = async () => {
     email: testUser.EMAIL,
     password: testUser.PASSWORD,
   });
-  // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-  headers = { headers: res.headers };
-  user = res.data.data;
+  store.dispatch(setHeaders(res.headers));
+  store.dispatch(setCurrentUser(res.data.data));
   return res;
 };
 
-const setHeaders = (newHeaders: ITokenHeaders) => {
-  const {
-    "content-type": contentType,
-    "access-token": accessToken,
-    client,
-    uid,
-  } = newHeaders;
-  headers = {
-    headers: {
-      "content-type": contentType,
-      "access-token": accessToken || headers.headers["access-token"],
-      client,
-      uid,
-    },
-  };
-};
 describe("/sign", () => {
   it("up", async () => {
     const res = await signUpTest();
@@ -93,58 +75,63 @@ describe("/sign", () => {
 
 describe(routes.MUSICS, () => {
   let music: IMusic;
+  const currentUser = getCurrentUser();
+  if (!currentUser) return;
   it("POST", async () => {
-    const res = await postMusic(
-      user.id,
-      {
-        title: "testMusic",
-        tab: "",
-        link_attributes: {
-          itunes: 0,
-          musixmatch: 0,
-          spotify: "spotify",
-        },
+    const res = await postMusic(getCurrentUser()?.id, {
+      title: "testMusic",
+      link_attributes: {
+        itunes: 0,
+        musixmatch: 0,
+        spotify: "spotify",
       },
-      headers
-    );
+    });
     music = res.data;
-    setHeaders(res.headers);
+    store.dispatch(setHeaders(res.headers));
     expect(res.status).toBe(201);
   });
   describe(routes.BOOKMARKS, () => {
     let bookmark: IMusicBookmark;
     it("POST", async () => {
-      const res = await postMusicBookmark(user.id, music.id, headers);
+      const res = await postMusicBookmark(currentUser.id, music.id);
       bookmark = res.data;
       setHeaders(res.headers);
       expect(res.status).toBe(201);
     });
     it("DELETE", async () => {
       const res = await deleteMusicBookmark(
-        user.id,
+        currentUser.id,
         music.id,
-        bookmark.id,
-        headers
+        bookmark.id
       );
       setHeaders(res.headers);
       expect(res.status).toBe(204);
     });
   });
+  describe(routes.ISSUES, () => {
+    it("POST", async () => {
+      const res = await postIssue(currentUser.id, music.id, {
+        title: "title",
+        description: "description",
+      });
+      setHeaders(res.headers);
+      expect(res.status).toBe(201);
+    });
+  });
   describe(routes.LINKS, () => {
     it("PATCH", async () => {
       const res = await patchMusicLink(
-        user.id,
+        currentUser.id,
         music.id,
         music.link.id,
-        { itunes: 0 },
-        headers
+        { itunes: 0 }
       );
       setHeaders(res.headers);
       expect(res.status).toBe(200);
     });
   });
   it("DELETE", async () => {
-    const res = await deleteMusic(user.id, music.id, headers);
+    const res = await deleteMusic(currentUser.id, music.id);
     setHeaders(res.headers);
     expect(res.status).toBe(204);
   });
@@ -152,31 +139,23 @@ describe(routes.MUSICS, () => {
 describe(routes.ALBUMS, () => {
   let album: IAlbum;
   it("POST", async () => {
-    const res = await postAlbum(
-      {
-        title: "testAlbum",
-        link_attributes: { itunes: 0, spotify: "spotify" },
-      },
-      headers
-    );
+    const res = await postAlbum({
+      title: "testAlbum",
+      link_attributes: { itunes: 0, spotify: "spotify" },
+    });
     album = res.data;
     setHeaders(res.headers);
     expect(res.status).toBe(201);
   });
   describe(routes.LINKS, () => {
     it("PATCH", async () => {
-      const res = await patchAlbumLink(
-        album.id,
-        album.link.id,
-        { itunes: 0 },
-        headers
-      );
+      const res = await patchAlbumLink(album.id, album.link.id, { itunes: 0 });
       setHeaders(res.headers);
       expect(res.status).toBe(200);
     });
   });
   it("DELETE", async () => {
-    const res = await deleteAlbum(album.id, headers);
+    const res = await deleteAlbum(album.id);
     setHeaders(res.headers);
     expect(res.status).toBe(204);
   });
@@ -185,18 +164,15 @@ describe(routes.ALBUMS, () => {
 describe(routes.BANDS, () => {
   let band: IBand;
   it("POST", async () => {
-    const res = await postBand(
-      {
-        name: "testBand",
-        link_attributes: {
-          itunes: 0,
-          spotify: "spotify",
-          wikipedia: 0,
-          twitter: "string",
-        },
+    const res = await postBand({
+      name: "testBand",
+      link_attributes: {
+        itunes: 0,
+        spotify: "spotify",
+        wikipedia: 0,
+        twitter: "string",
       },
-      headers
-    );
+    });
     band = res.data;
     setHeaders(res.headers);
     expect(res.status).toBe(201);
@@ -204,31 +180,26 @@ describe(routes.BANDS, () => {
   describe(routes.BOOKMARKS, () => {
     let bookmark: IBandBookmark;
     it("POST", async () => {
-      const res = await postBandBookmark(band.id, headers);
+      const res = await postBandBookmark(band.id);
       bookmark = res.data;
       setHeaders(res.headers);
       expect(res.status).toBe(201);
     });
     it("DELETE", async () => {
-      const res = await deleteBandBookmark(band.id, bookmark.id, headers);
+      const res = await deleteBandBookmark(band.id, bookmark.id);
       setHeaders(res.headers);
       expect(res.status).toBe(204);
     });
   });
   describe(routes.LINKS, () => {
     it("PATCH", async () => {
-      const res = await patchBandLink(
-        band.id,
-        band.link.id,
-        { itunes: 0 },
-        headers
-      );
+      const res = await patchBandLink(band.id, band.link.id, { itunes: 0 });
       setHeaders(res.headers);
       expect(res.status).toBe(200);
     });
   });
   it("DELETE", async () => {
-    const res = await deleteBand(band.id, headers);
+    const res = await deleteBand(band.id);
     setHeaders(res.headers);
     expect(res.status).toBe(204);
   });
@@ -236,18 +207,15 @@ describe(routes.BANDS, () => {
 describe(routes.ARTISTS, () => {
   let artist: IArtist;
   it("POST", async () => {
-    const res = await postArtist(
-      {
-        name: "testArtist",
-        link_attributes: {
-          itunes: 0,
-          spotify: "spotify",
-          wikipedia: 0,
-          twitter: "string",
-        },
+    const res = await postArtist({
+      name: "testArtist",
+      link_attributes: {
+        itunes: 0,
+        spotify: "spotify",
+        wikipedia: 0,
+        twitter: "string",
       },
-      headers
-    );
+    });
     artist = res.data;
     setHeaders(res.headers);
     expect(res.status).toBe(201);
@@ -255,39 +223,38 @@ describe(routes.ARTISTS, () => {
   describe(routes.BOOKMARKS, () => {
     let bookmark: IArtistBookmark;
     it("POST", async () => {
-      const res = await postArtistBookmark(artist.id, headers);
+      const res = await postArtistBookmark(artist.id);
       bookmark = res.data;
       setHeaders(res.headers);
       expect(res.status).toBe(201);
     });
     it("DELETE", async () => {
-      const res = await deleteArtistBookmark(artist.id, bookmark.id, headers);
+      const res = await deleteArtistBookmark(artist.id, bookmark.id);
       setHeaders(res.headers);
       expect(res.status).toBe(204);
     });
   });
   describe(routes.LINKS, () => {
     it("PATCH", async () => {
-      const res = await patchArtistLink(
-        artist.id,
-        artist.link.id,
-        { itunes: 0 },
-        headers
-      );
+      const res = await patchArtistLink(artist.id, artist.link.id, {
+        itunes: 0,
+      });
       setHeaders(res.headers);
       expect(res.status).toBe(200);
     });
   });
   it("DELETE", async () => {
-    const res = await deleteArtist(artist.id, headers);
+    const res = await deleteArtist(artist.id);
     setHeaders(res.headers);
     expect(res.status).toBe(204);
   });
 });
 
 describe("/user", () => {
+  const currentUser = getCurrentUser();
+  if (!currentUser) return;
   it("DELETE", async () => {
-    const res = await deleteUser(user.id);
+    const res = await deleteUser(currentUser.id);
     expect(res.status).toBe(204);
   });
 });
